+++ src/main/java/techcourse/myblog/domain/repository/ArticleRepository.java
@@ -1,7 +1,8 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import techcourse.myblog.domain.Article;
 
 @Repository
 public interface ArticleRepository extends JpaRepository<Article, Long> {
