+++ src/main/java/techcourse/myblog/domain/repository/UserRepository.java
@@ -1,7 +1,8 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import techcourse.myblog.domain.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
