+++ src/test/java/techcourse/myblog/domain/UserRepositoryTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import techcourse.myblog.domain.repository.UserRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
